Created the mimic mechanic for the prey
I created the mimic mechanic so the prey can copy the sprite of other objects which will change it's Image, Color, and Size. I added a death function as well but haven't tested it to see if it works yet. I also created a new scene to test the mimic mechanic.